requirements.txt
requests
pyserial
RPi.GPIO
